sdk/rust: consolidate tests and extract shared helpers
Delete three redundant test files (codegen.rs, export_macros.rs,
provider_server.rs) whose assertions were compile-only checks or
duplicates of behavior already covered by the transport integration
tests. The one unique test from provider_server.rs that exercises
error paths (unknown operation 404, decode error, handler error,
panic recovery) is relocated into router.rs where it belongs.

Extract duplicated test utilities (EnvGuard, struct_from_json,
json_to_prost, wait_for_socket, temp_socket, env_lock) into a
shared helpers.rs module, replacing the copy-pasted definitions in
transport.rs and component_transport.rs. This also fixes a latent
race condition in transport.rs by adding the env_lock that
component_transport.rs already had. Net result: 18 tests across 8
files reduced to 10 tests across 5 files (plus the shared helpers).
